+++ app/main.rb
@@ -1,7 +1,8 @@
 # main.rb
 
 require_relative 'person'
-require_relative 'decorators'
+require_relative 'capitalize_decorator'
+require_relative 'trimmer_decorator'
 
 person = Person.new(22, 'maximilianus')
 puts person.correct_name
