+++ app/student.rb
@@ -1,7 +1,7 @@
 # Represents a teacher, inheriting from the Person class.
 class Student < Person
-  def initialize(id, age, classroom, name = 'Unknown', parent_permission: true)
-    super(id, age, name, parent_permission)
+  def initialize(age, classroom, name = 'Unknown', parent_permission: true)
+    super(age, name, parent_permission)
     @classroom = classroom
   end
