+++ app/teacher.rb
@@ -1,7 +1,7 @@
 # Represents a teacher, inheriting from the Person class.
 class Teacher < Person
-  def initialize(id, age, specialization, name = 'Unknown', parent_permission: true)
-    super(id, age, name, parent_permission)
+  def initialize(age, specialization, name = 'Unknown', parent_permission: true)
+    super(age, name, parent_permission)
     @specialization = specialization
   end
